add basic grades dashboard
load the page with a specific user ID to show the latest
submissions for each assignment.
Down the line: plan on adding a 'view submission history' button
for each individual assignment so the students can see all of their
submissions.

Sall note on the time zones: datetime.fromtimestamp seems to
automatically convert the datetime to the timezone of the host machine.
May want to be more explicit about this in case the server is located
somewhere else.